+++ service/movieApi/src/action/getHandlerLatestFileList.js
@@ -1,6 +1,7 @@
 export const getHandlerLatestFileList = ({ handleLatestFileList }) => {
   return async (req, res) => {
-    const handleResult = handleLatestFileList()
+    const { requestId } = req.query
+    const handleResult = handleLatestFileList({ requestId })
 
     res.json({ result: handleResult })
   }
