+++ service/movieApi/src/core/handleLatestFileList.js
@@ -1,10 +1,10 @@
 import { mod, store } from './init.js'
 export default {}
 
-export const handleLatestFileList = () => {
+export const handleLatestFileList = ({ requestId }) => {
   const MOVIE_DIR_PATH = mod.setting.getValue('path.MOVIE_DIR_PATH') 
-  const dirPath = MOVIE_DIR_PATH
-  const fileDirList = mod.input.getFileDirListWithLatestTitle({ dirPath })
+  const dirPath = `${MOVIE_DIR_PATH}${requestId}/`
+  const fileDirList = mod.input.getLatestFileList({ dirPath })
   const handleResult = { result: { fileDirList } }
   return handleResult
 }
