+++ service/movieApi/src/input.js
@@ -2,13 +2,13 @@ import { init } from './input/init.js'
 import { getFileDirList } from './input/getFileDirList.js'
 import { getFileDirListWithTitle } from './input/getFileDirListWithTitle.js'
 import { getFileContent } from './input/getFileContent.js'
-import { getFileDirListWithLatestTitle } from './input/getFileDirListWithLatestTitle.js'
+import { getLatestFileList } from './input/getLatestFileList.js'
 
 export default {
   init,
   getFileDirList,
   getFileDirListWithTitle,
   getFileContent,
-  getFileDirListWithLatestTitle,
+  getLatestFileList,
 }
